Исправлена логика перехода в полноэкранный режим. Добавлены стили компонента 'PayButton'
