+++ project/src/utils.ts
@@ -106,4 +106,29 @@ const getSetOfAvailableGenres = (filmsToFilter : Film[]) : Set<string> => {
   return genres;
 };
 
-export {lightenDarkenColor, getRatingLevel, getRandomPositiveNumber, getUniqueRandomNumbers, getFilmRunTime, getSetOfAvailableGenres};
+const tarsformTimeFormat = (value : number) : string => {
+  let hours = 0;
+  let minutes = 0;
+  let seconds = 0;
+  let result = '';
+
+  if(value > 3599){
+    hours = Math.floor(value / 3600);
+    minutes = Math.floor((value - hours * 3600) / 60);
+    value = value - (60 * minutes) - (3600 * hours);
+    seconds = value;
+
+    result = `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+  }
+  else{
+    minutes = Math.floor(value / 60);
+    value = value - (60 * minutes);
+    seconds = value;
+
+    result = `${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+  }
+
+  return result;
+};
+
+export {lightenDarkenColor, getRatingLevel, getRandomPositiveNumber, getUniqueRandomNumbers, getFilmRunTime, getSetOfAvailableGenres, tarsformTimeFormat};
